Improving shapes

+++ wirez-sets/wirez-bpmn/wirez-bpmn-api/src/main/java/org/wirez/bpmn/api/EndNoneEvent.java
@@ -40,7 +40,7 @@ public class EndNoneEvent extends BaseDefinition implements BPMNDefinition {
 
     public static final String ID = "EndNoneEvent";
     public static final String COLOR = "#ff0000";
-    public static final Integer RADIUS = 25;
+    public static final Double RADIUS = 25d;
     
     @PropertySet
     private BPMNGeneral general;
